app/requirements.txt
flask==3.0.0
azure-storage-blob==12.19.0
requests==2.31.0